chore: relocate YearSelector to common folder
