+++ src/components/common/YearSelector.tsx
@@ -1,8 +1,8 @@
 import { useRef, useState } from "react";
-import { useSelection } from "../../../context/useSelection";
-import { yearList } from "../../../data/list";
-import useClickOutside from "../../../hooks/useClickOutside";
-import { toggleDropdown } from "../../../utils/toggleDropdown";
+import { useSelection } from "../../context/useSelection";
+import { yearList } from "../../data/list";
+import useClickOutside from "../../hooks/useClickOutside";
+import { toggleDropdown } from "../../utils/toggleDropdown";
 
 type Mode = "map" | "selector";
